--- rl_agents/envs/__init__.py
@@ -1,2 +1,0 @@
-# from rl_agents.envs.molten_drop import MoltenDropEnv
-# from rl_agents.envs.sfk_fight import SFKFightEnv
