+++ rl_agents/state/state.py
@@ -1,16 +1,10 @@
 import subprocess
 import uuid
 import skimage.io
-# import torch
 from typing import Dict, Union
 import numpy as np
-# import torchvision.transforms as T
 
 from rl_agents.state.overlay import OverlayHealth, OverlayMana
-
-
-# comp = T.Compose([T.ToPILImage(), T.ToTensor()])
-# totensor = T.Compose([T.ToTensor()])
 
 
 def read_image(path: str) -> np.ndarray:
@@ -46,26 +40,8 @@ def cut_overlay(image: np.ndarray) -> np.ndarray:
 
 
 def get_state():
-    # print("NEW IMAGE STATE..")
     image_path = take_screenshot()
     screen = read_image(image_path)
     screen = cut_overlay(screen)
     overlay_info = parse_overlay(screen)
-    # print(screen.shape)
-    # exit()
-    # screen = screen.transpose((2, 0, 1))
-
-    # screen = np.ascontiguousarray(screen, dtype=np.float32)
-
-    # screen = torch.from_numpy(screen)
-    # screen = comp(screen)
-    # print(screen.shape)
-
-    # screen = totensor(screen)
-
-    # print(screen.shape)
-
-    # screen = screen.unsqueeze(0)
-
-    # print(screen.shape)
     return screen, overlay_info
