+++ rl_agents/train.py
@@ -1,166 +1,144 @@
-import gym
-import math
-import random
-import torch
-from itertools import count
+import pyvirtualdisplay
+from tf_agents.environments import suite_gym, tf_py_environment
+from tf_agents.policies import random_tf_policy
+from tf_agents.specs import tensor_spec
+from tf_agents.trajectories import trajectory
+from tf_agents.utils import common
 import matplotlib.pyplot as plt
-import torch.nn as nn
-import torch.optim as optim
 
-from rl_agents.dqn import DQN
-from rl_agents.input_extraction import get_screen
-from rl_agents.replay_memory import ReplayMemory, Transition
-
-
-BATCH_SIZE = 128
-GAMMA = 0.999
-EPS_START = 0.9
-EPS_END = 0.05
-EPS_DECAY = 200
-TARGET_UPDATE = 10
-
-device = torch.device("cpu")
+from rl_agents.envs.onyxia import OnyxiaEnv
+from rl_agents.models.dqn.dqn import SimpleDQN
 
 
 class TrainingManager:
-    def __init__(self, gym_env: str, channels: int, unwrapped: bool = False):
+    def __init__(
+        self, 
+        env_name: str, 
+        channels: int,
+        num_iterations: int,
+        collect_steps_per_iteration: int,
+        learning_rate: float,
+        replay_buffer_max_len: int,
+        batch_size: int,
+        num_eval_episodes: int,
+        initial_collect_steps: int,
+        log_interval: int,
+        eval_interval: int,
+        episode_max_steps: int,
+    ):
         """
         Creates a `TrainingManager` object, used for RL training in a given
         gym environment.
 
         :param gym_env: name of the gym environment to use
         :param channels: number of channels to use in images (1=grayscale, 3=rgb)
-        :param unwrapped: use unwrapped gym environment (default=False)
         """
-        self.env = gym.make(gym_env)
-        if unwrapped:
-            self.env = self.env.unwrapped
-        self.env.reset()
-        init_screen = get_screen(self.env)
-        _, _, screen_height, screen_width = init_screen.shape
-
-        # Get number of actions from gym action space
-        self.n_actions = self.env.action_space.n
-        self.policy_net = DQN(screen_height, screen_width, self.n_actions, 3).to(device)
-        self.target_net = DQN(screen_height, screen_width, self.n_actions, 3).to(device)
-        self.target_net.load_state_dict(self.policy_net.state_dict())
-        self.target_net.eval()
-
-        self.optimizer = optim.RMSprop(self.policy_net.parameters())
-        self.memory = ReplayMemory(10000)
-
-        self.steps_done = 0
-        self.episode_durations = []
-    
-    def select_action(self, state):
-        sample = random.random()
-        eps_threshold = EPS_END + (EPS_START - EPS_END) * \
-            math.exp(-1. * self.steps_done / EPS_DECAY)
-        self.steps_done += 1
-        if sample > eps_threshold:
-            with torch.no_grad():
-                # t.max(1) will return largest column value of each row.
-                # second column on max result is index of where max element was
-                # found, so we pick action with the larger expected reward.
-                return self.policy_net(state).max(1)[1].view(1, 1)
+        self.env_name = env_name
+        if env_name == 'WoW':
+            train_py_env = OnyxiaEnv()
+            eval_py_env = train_py_env
         else:
-            return torch.tensor([[random.randrange(self.n_actions)]], device=device, dtype=torch.long)
+            train_py_env = suite_gym.load(env_name)
+            eval_py_env = suite_gym.load(env_name)
+            # is this doing anything
+            display = pyvirtualdisplay.Display(visible=0, size=(1400, 900)).start()
+        print(f"Observation spec: {train_py_env.time_step_spec().observation}")
+        print(f"Reward Spec: {train_py_env.time_step_spec().reward}")
+        print(f"Action space: {train_py_env.action_spec()}")
 
-    def optimize_model(self):
-        if len(self.memory) < BATCH_SIZE:
-            return
-        transitions = self.memory.sample(BATCH_SIZE)
-        # Transpose the batch (see https://stackoverflow.com/a/19343/3343043 for
-        # detailed explanation). This converts batch-array of Transitions
-        # to Transition of batch-arrays.
-        batch = Transition(*zip(*transitions))
+        self.train_env = tf_py_environment.TFPyEnvironment(train_py_env)
+        self.eval_env = tf_py_environment.TFPyEnvironment(eval_py_env)
 
-        # Compute a mask of non-final states and concatenate the batch elements
-        # (a final state would've been the one after which simulation ended)
-        non_final_mask = torch.tensor(tuple(map(lambda s: s is not None,
-                                            batch.next_state)), device=device, dtype=torch.bool)
-        non_final_next_states = torch.cat([s for s in batch.next_state
-                                                    if s is not None])
-        state_batch = torch.cat(batch.state)
-        action_batch = torch.cat(batch.action)
-        reward_batch = torch.cat(batch.reward)
+        self.num_iterations = num_iterations
+        self.collect_steps_per_iteration = collect_steps_per_iteration
+        self.num_eval_episodes = num_eval_episodes
+        self.log_interval = log_interval
+        self.eval_interval = eval_interval
+        self.episode_max_steps = episode_max_steps
 
-        # Compute Q(s_t, a) - the model computes Q(s_t), then we select the
-        # columns of actions taken. These are the actions which would've been taken
-        # for each batch state according to policy_net
-        state_action_values = self.policy_net(state_batch).gather(1, action_batch)
+        action_tensor_spec = tensor_spec.from_spec(train_py_env.action_spec())
+        num_actions = action_tensor_spec.maximum - action_tensor_spec.minimum + 1
+        print(f"Num actions = {num_actions}")
+        self.dqn = SimpleDQN((100, 50), num_actions, learning_rate, self.train_env, replay_buffer_max_len)
+        random_policy = random_tf_policy.RandomTFPolicy(self.train_env.time_step_spec(), self.train_env.action_spec())
+        self.train_env.reset()
+        print("Collecting initial steps")
+        self.collect_data(initial_collect_steps)
+        # Dataset generates trajectories with shape [Bx2x...]
+        dataset = self.dqn.replay_buffer.as_dataset(
+            num_parallel_calls=3, 
+            sample_batch_size=batch_size, 
+            num_steps=2
+        ).prefetch(3)
 
-        # Compute V(s_{t+1}) for all next states.
-        # Expected values of actions for non_final_next_states are computed based
-        # on the "older" target_net; selecting their best reward with max(1)[0].
-        # This is merged based on the mask, such that we'll have either the expected
-        # state value or 0 in case the state was final.
-        next_state_values = torch.zeros(BATCH_SIZE, device=device)
-        next_state_values[non_final_mask] = self.target_net(non_final_next_states).max(1)[0].detach()
-        # Compute the expected Q values
-        expected_state_action_values = (next_state_values * GAMMA) + reward_batch
+        self.iterator = iter(dataset)
+        # print(f"Baseline: {self.compute_avg_return(random_policy)}")
+    
+    def collect_step(self):
+        time_step = self.train_env.current_time_step()
+        action_step = self.dqn.agent.policy.action(time_step)
+        next_time_step = self.train_env.step(action_step.action)
+        traj = trajectory.from_transition(time_step, action_step, next_time_step)
+         # Add trajectory to the replay buffer
+        self.dqn.replay_buffer.add_batch(traj)
+        return next_time_step.is_last()
 
-        # Compute Huber loss
-        criterion = nn.SmoothL1Loss()
-        loss = criterion(state_action_values, expected_state_action_values.unsqueeze(1))
+    def collect_data(self, steps: int) -> bool:
+        is_last = False
+        for _ in range(steps):
+            is_last = self.collect_step()
+            if is_last:
+                break
+        return is_last
+    
+    def compute_avg_return(self, policy):
+        print("Computing average return")
+        total_return = 0.0
+        for _ in range(self.num_eval_episodes):
+            time_step = self.eval_env.reset()
+            episode_return = 0.0
 
-        # Optimize the model
-        self.optimizer.zero_grad()
-        loss.backward()
-        for param in self.policy_net.parameters():
-            param.grad.data.clamp_(-1, 1)
-        self.optimizer.step()
+            while not time_step.is_last():
+                action_step = policy.action(time_step)
+                time_step = self.eval_env.step(action_step.action)
+                episode_return += time_step.reward
+            total_return += episode_return
+
+        avg_return = total_return / self.num_eval_episodes
+        return avg_return.numpy()[0]
+
 
     def train(self):
-        num_episodes = 1000
-        for i_episode in range(num_episodes):
-            # Initialize the environment and state
-            self.env.reset()
-            last_screen = get_screen(self.env)
-            current_screen = get_screen(self.env)
-            state = current_screen - last_screen
-            for t in count():
-                # Select and perform an action
-                action = self.select_action(state)
-                _, reward, done, _ = self.env.step(action.item())
-                reward = torch.tensor([reward], device=device)
+        # (Optional) Optimize by wrapping some of the code in a graph using TF function.
+        self.dqn.agent.train = common.function(self.dqn.agent.train)
+        # Reset the train step
+        self.dqn.agent.train_step_counter.assign(0)
+        # Evaluate the agent's policy once before training.
+        avg_return = self.compute_avg_return(self.dqn.agent.policy)
+        returns = [avg_return]
 
-                # Observe new state
-                last_screen = current_screen
-                current_screen = get_screen(self.env)
-                if not done:
-                    next_state = current_screen - last_screen
-                else:
-                    next_state = None
-
-                # Store the transition in memory
-                self.memory.push(state, action, next_state, reward)
-
-                # Move to the next state
-                state = next_state
-
-                # Perform one step of the optimization (on the policy network)
-                self.optimize_model()
-                if done:
-                    self.episode_durations.append(t + 1)
-                    self.plot_durations()
+        print("Training agent")
+        for epi_i in range(self.num_iterations):
+            print(f"Resetting for episode {epi_i}")
+            self.train_env.reset()
+            for _ in range(self.episode_max_steps):
+                # Collect step using collect_policy and save to the replay buffer.
+                is_last = self.collect_data(self.collect_steps_per_iteration)
+                # Sample a batch of data from the buffer and update the agent's network.
+                experience, unused_info = next(self.iterator)
+                train_loss = self.dqn.agent.train(experience).loss
+                step = self.dqn.agent.train_step_counter.numpy()
+                if step % self.log_interval == 0:
+                    print('episode = {2}: step = {0}: loss = {1}'.format(step, train_loss, epi_i))
+                if is_last:
                     break
-            # Update the target network, copying all weights and biases in DQN
-            if i_episode % TARGET_UPDATE == 0:
-                self.target_net.load_state_dict(self.policy_net.state_dict())
-    
-    def plot_durations(self):
-        plt.figure(2)
-        plt.clf()
-        durations_t = torch.tensor(self.episode_durations, dtype=torch.float)
-        plt.title('Training...')
-        plt.xlabel('Episode')
-        plt.ylabel('Duration')
-        plt.plot(durations_t.numpy())
-        # Take 100 episode averages and plot them too
-        if len(durations_t) >= 100:
-            means = durations_t.unfold(0, 100, 1).mean(1).view(-1)
-            means = torch.cat((torch.zeros(99), means))
-            plt.plot(means.numpy())
-
-        plt.pause(0.001)  # pause a bit so that plots are updated
+            if epi_i != 0 and epi_i % self.eval_interval == 0:
+                avg_return = self.compute_avg_return(self.dqn.agent.policy)
+                step = self.dqn.agent.train_step_counter.numpy()
+                print('step = {0}: Average Return = {1}'.format(step, avg_return))
+                returns.append(avg_return)
+                iterations = range(0, epi_i + 1, self.eval_interval)
+                plt.plot(iterations, returns)
+                plt.ylabel('Average Return')
+                plt.xlabel('Iterations')
+                plt.savefig("output/resultsnewish.png")
